Added admin option to keep images in the data/tmp folder. 
Added admin option to enable EXIF data preservation, default disable

+++ api/applyEffects.php
@@ -99,7 +99,7 @@ imagedestroy($thumbResource);
 if ($imageModified || $config['jpeg_quality_image'] !== -1) {
     imagejpeg($imageResource, $filename_photo, $config['jpeg_quality_image']);
     // preserve jpeg meta data
-    if ($config['exiftool']['cmd']) {
+    if ($config['preserve_exif_data'] && $config['exiftool']['cmd']) {
         $cmd = sprintf($config['exiftool']['cmd'], $filename_tmp, $filename_photo);
         exec($cmd, $output, $returnValue);
         if ($returnValue) {
@@ -114,7 +114,11 @@ if ($imageModified || $config['jpeg_quality_image'] !== -1) {
 } else {
     copy ( $filename_tmp, $filename_photo );
 }
-unlink ($filename_tmp);
+
+if (!$config['keep_images']) {
+    unlink($filename_tmp);
+}
+
 imagedestroy($imageResource);
 
 // insert into database
